tests/submission/W3C/domhtml/HTMLIFrameElement10.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IFRAME</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/dimension-attributes.html#dom-dim-width" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<iframe src="/support/html/sample.html" width='550'>IFRAME1</iframe>

<script>
var domtest = async_test("The width attribute specifies the frame width.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vwidth;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("iframe");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vwidth = testNode.width;

      assert_equals(vwidth,"550");
       
    domtest.done();
}));

</script>
<p>The width attribute specifies the frame width. 

    Retrieve the width attribute of the first IFRAME element and examine
    it's value.</p>
<div id="log"></div>
</body></html>
